% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bioscreen.R
\name{bioscreen_example}
\alias{bioscreen_example}
\title{get a gp1 example data file}
\usage{
bioscreen_example(file)
}
\arguments{
\item{file}{\code{<chr>} name of the file}
}
\description{
get a gp1 example data file
}
\examples{
    bioscreen_example('data.csv')
    bioscreen_example('data-2.csv')
}
