% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bioscreen.R
\name{infer_header}
\alias{infer_header}
\title{infer the header parameters of bioscreen file}
\usage{
infer_header(file)
}
\arguments{
\item{file}{<chr> path to file}
}
\description{
infer the header parameters of bioscreen file
}
\keyword{internal}
