% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bioscreen.R
\name{time_to_runtime}
\alias{time_to_runtime}
\title{convert the raw bioscreen time format to runtime in seconds.}
\usage{
time_to_runtime(x)
}
\arguments{
\item{x}{<chr> string of the form 'hours:minutes:seconds' where minutes and
seconds are represented by 1-2 digits while hours (elapsed) can be >=1}
}
\description{
convert the raw bioscreen time format to runtime in seconds.
}
\keyword{internal}
