% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bioscreen.R
\name{translate_well_int}
\alias{translate_well_int}
\title{translate column names to plate and 'a01' style well notation}
\usage{
translate_well_int(v)
}
\arguments{
\item{x}{<chr> column name to translate
translate_well_int(c(1,10,11,99,100,101,200,201, "well_1", "well_101"))}
}
\description{
translate column names to plate and 'a01' style well notation
}
\keyword{internal}
